requirements.txt
alembic==1.9.4
anyio==3.6.2
attrs==22.2.0
autopep8==2.0.2
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
cssselect==1.2.0
cssutils==2.6.0
dnspython==2.3.0
ecdsa==0.18.0
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.2.0
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
iniconfig==2.0.0
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
premailer==3.10.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycodestyle==2.10.0
pydantic==1.10.5
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
tenacity==8.2.2
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.21.0
zipp==3.15.0
